// AngularERP/src/app/models/sales-purchase/customer.model.ts
export class Customer {
    accountId: number;
    accountName: string;
    aliasName: string;
    formattedName: string;
    deliveryAddress: string;
    locationId: number;
    locationName: string;
    pertyCurrentBalance: number;
    commissionTypeID: number;
    commissionTypeName: string;
    billContactNo:string;
    openingBalance:number;
    creditLimit:number;
}